Change float to decimal type

+++ masteraula/questions/models.py
@@ -605,7 +605,7 @@ class FaqQuestion(models.Model):
 class StudentAnswer(models.Model):
     answer_alternative =  models.ForeignKey(Alternative, on_delete=models.CASCADE, null=True, blank=True)
     answer_text = models.TextField(null=True, blank=True)
-    score_answer = models.FloatField(null=True, blank=True)
+    score_answer = models.DecimalField(null=True, blank=True, max_digits=5, decimal_places=2)
     student_answer = models.ForeignKey('Result', related_name='student_answer', on_delete=models.CASCADE, null=True, blank=True)
     student_question = models.ForeignKey('DocumentQuestionOnline', related_name='student_question', on_delete=models.CASCADE)
 
@@ -643,7 +643,7 @@ class Result(models.Model):
     student_levels = models.CharField(max_length=200)
     start = models.DateTimeField()
     finish = models.DateTimeField()
-    total_score =  models.FloatField(null=True, blank=True)
+    total_score =  models.DecimalField(null=True, blank=True, max_digits=5, decimal_places=2)
 
     objects = ResultManager()
 
@@ -731,7 +731,7 @@ class DocumentQuestionOnlineManager(models.Manager):
 class DocumentQuestionOnline(models.Model):
     question = models.ForeignKey(Question, on_delete=models.CASCADE)
     document = models.ForeignKey(DocumentOnline, on_delete=models.CASCADE)
-    score = models.FloatField(null=True, blank=True)
+    score = models.DecimalField(null=True, blank=True, max_digits=5, decimal_places=2)
 
     objects = DocumentQuestionOnlineManager()
